rename user_constant.py as consts.py in console
