+++ src/qianfan/resources/console/consts.py
